# swagger/swagger.yaml
definitions:
...
  models.User:
    properties:
      created_at:
        type: string
      disabled:
        description: 是否启用
        type: boolean
      group_names:
        type: string
      id:
        type: integer
      password:
        type: string
      salt:
        type: string
      source:
        description: 来源，如：db, ldap_config.json, oauth
        type: string
      two_fa_app_name:
        description: 2FA应用名称，用于提醒用户使用的是哪个软件
        type: string
      two_fa_backup_codes:
        description: 备用恢复码，逗号分隔
        type: string
      two_fa_enabled:
        description: 是否启用2FA
        type: boolean
      two_fa_secret:
        description: 2FA密钥
        type: string
      two_fa_type:
        description: 2FA类型：如 'totp', 'sms', 'email'
        type: string
      updated_at:
        description: Automatically managed by GORM for update time
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: 请输入以 `Bearer ` 开头的 Token，例：Bearer xxxxxxxx。未列出接口请参考前端调用方法。Token在个人中心-API密钥菜单下申请。
  title: k8m API
  version: "1.0"
paths:
...
  /admin/user/2fa/disable/{id}:
    post:
      description: 禁用指定用户的二步验证
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 禁用用户2FA
  /admin/user/delete/{ids}:
    post:
      description: 根据ID批量删除用户
      parameters:
      - description: 用户ID，多个用逗号分隔
        in: path
        name: ids
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除用户
  /admin/user/list:
    get:
      description: 获取所有用户信息
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerAuth: []
      summary: 获取用户列表
  /admin/user/option_list:
    get:
      description: 获取用户选项列表，用于下拉选择
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户选项列表
  /admin/user/save:
    post:
      consumes:
      - application/json
      description: 新增或更新用户信息
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 保存用户
  /admin/user/save/id/{id}/status/{disabled}:
    post:
      description: 根据ID快速更新用户启用/禁用状态
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态，例如：true、false
        in: path
        name: disabled
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 快速更新用户状态
  /admin/user/update_psw/{id}:
    post:
      consumes:
      - application/json
      description: 根据ID更新用户密码
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 新密码信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新用户密码
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/annotations/ns/{ns}/name/{name}:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 列出资源注解
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/update_annotations/ns/{ns}/name/{name}:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      - description: 注解键值对
        in: body
        name: annotations
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新资源注解
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/update_labels/ns/{ns}/name/{name}:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      - description: 标签键值对
        in: body
        name: labels
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新资源标签
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"